+++ gmo_fx/api/klines.py
@@ -4,7 +4,7 @@ from enum import Enum
 from typing import Literal
 from requests import get, Response
 from gmo_fx.api_base import PublicApiBase
-from gmo_fx.response import Response as ResponseBase
+from api.response import Response as ResponseBase
 from gmo_fx.symbols import Symbol
 from gmo_fx.urls import BASE_URL_PUBLIC
